[C-IRIS] move c_iris_collision_geometry out of dev folder.
