+++ geometry/optimization/BUILD.bazel
@@ -14,6 +14,7 @@ drake_cc_package_library(
     name = "optimization",
     visibility = ["//visibility:public"],
     deps = [
+        ":c_iris_collision_geometry",
         ":convex_set",
         ":graph_of_convex_sets",
         ":iris",
@@ -56,6 +57,19 @@ drake_cc_library(
         "//solvers:scs_solver",
         "//solvers:solve",
         "@qhull_internal//:qhull",
+    ],
+)
+
+drake_cc_library(
+    name = "c_iris_collision_geometry",
+    srcs = ["c_iris_collision_geometry.cc"],
+    hdrs = ["c_iris_collision_geometry.h"],
+    deps = [
+        "//geometry:geometry_ids",
+        "//geometry:shape_specification",
+        "//geometry/optimization:convex_set",
+        "//multibody/rational:rational_forward_kinematics",
+        "//multibody/tree:multibody_tree_indexes",
     ],
 )
 
@@ -115,6 +129,32 @@ drake_cc_googletest(
         ":convex_set",
         "//common/test_utilities:eigen_matrix_compare",
         "//solvers:clp_solver",
+    ],
+)
+
+drake_cc_library(
+    name = "c_iris_test_utilities",
+    testonly = 1,
+    srcs = ["test/c_iris_test_utilities.cc"],
+    hdrs = ["test/c_iris_test_utilities.h"],
+    data = [
+        "test/convex.mtl",
+        "test/convex.obj",
+    ],
+    deps = [
+        "//geometry:scene_graph",
+        "//multibody/plant",
+        "//systems/framework:diagram_builder",
+        "@gtest//:without_main",
+    ],
+)
+
+drake_cc_googletest(
+    name = "c_iris_collision_geometry_test",
+    deps = [
+        ":c_iris_collision_geometry",
+        ":c_iris_test_utilities",
+        "//common/test_utilities:symbolic_test_util",
     ],
 )
