+++ geometry/optimization/dev/BUILD.bazel
@@ -13,25 +13,12 @@ load(
 load("//tools/lint:lint.bzl", "add_lint_tests")
 
 drake_cc_library(
-    name = "c_iris_collision_geometry",
-    srcs = ["c_iris_collision_geometry.cc"],
-    hdrs = ["c_iris_collision_geometry.h"],
-    deps = [
-        "//geometry:geometry_ids",
-        "//geometry:shape_specification",
-        "//geometry/optimization:convex_set",
-        "//multibody/rational:rational_forward_kinematics",
-        "//multibody/tree:multibody_tree_indexes",
-    ],
-)
-
-drake_cc_library(
     name = "c_iris_separating_plane",
     srcs = ["c_iris_separating_plane.cc"],
     hdrs = ["c_iris_separating_plane.h"],
     deps = [
-        ":c_iris_collision_geometry",
         "//common/symbolic:polynomial",
+        "//geometry/optimization:c_iris_collision_geometry",
     ],
 )
 
@@ -40,9 +27,9 @@ drake_cc_library(
     srcs = ["cspace_free_polytope.cc"],
     hdrs = ["cspace_free_polytope.h"],
     deps = [
-        ":c_iris_collision_geometry",
         ":c_iris_separating_plane",
         "//common/symbolic:monomial_util",
+        "//geometry/optimization:c_iris_collision_geometry",
         "//solvers:choose_best_solver",
         "//solvers:mathematical_program",
         "//solvers:mosek_solver",
@@ -51,31 +38,16 @@ drake_cc_library(
 )
 
 # test
-
 drake_cc_library(
     name = "c_iris_test_utilities",
     testonly = 1,
     srcs = ["test/c_iris_test_utilities.cc"],
     hdrs = ["test/c_iris_test_utilities.h"],
-    data = [
-        "test/convex.mtl",
-        "test/convex.obj",
-    ],
     deps = [
         ":cspace_free_polytope",
         "//geometry:scene_graph",
         "//multibody/plant",
         "//systems/framework:diagram_builder",
-        "@gtest//:without_main",
-    ],
-)
-
-drake_cc_googletest(
-    name = "c_iris_collision_geometry_test",
-    deps = [
-        ":c_iris_collision_geometry",
-        ":c_iris_test_utilities",
-        "//common/test_utilities:symbolic_test_util",
     ],
 )
 
@@ -94,6 +66,7 @@ drake_cc_googletest(
         ":cspace_free_polytope",
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:symbolic_test_util",
+        "//geometry/optimization:c_iris_test_utilities",
     ],
 )
 
@@ -109,6 +82,7 @@ drake_cc_googletest(
         ":cspace_free_polytope",
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:symbolic_test_util",
+        "//geometry/optimization:c_iris_test_utilities",
         "//solvers:mosek_solver",
         "//solvers:sos_basis_generator",
     ],
