+++ geometry/optimization/dev/test/cspace_free_polytope_with_mosek_test.cc
@@ -8,6 +8,7 @@
 #include "drake/geometry/geometry_ids.h"
 #include "drake/geometry/optimization/dev/cspace_free_polytope.h"  // NOLINT
 #include "drake/geometry/optimization/dev/test/c_iris_test_utilities.h"
+#include "drake/geometry/optimization/test/c_iris_test_utilities.h"
 #include "drake/multibody/rational/rational_forward_kinematics.h"
 #include "drake/multibody/rational/rational_forward_kinematics_internal.h"
 #include "drake/solvers/common_solver_option.h"
